Used package template (flutter create -t package .) Made untis.dart a library Fixed warning about not returning value
