+++ pubspec.yaml
@@ -1,0 +1,29 @@
+name: dart_webuntis
+description: Asynchrous WebUntis API wrapper written in Dart
+version: 1.0.0
+homepage: https://github.com/IsAvaible/dart-webuntis
+
+# The following line prevents the package from being accidentally published to
+# pub.dev using `flutter pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+
+environment:
+  sdk: '>=3.2.0-42.2.beta <4.0.0'
+
+dependencies:
+  flutter:
+    sdk: flutter
+    
+  http: ^0.13.4
+  string_similarity: ^2.0.0
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+
+  # The "flutter_lints" package below contains a set of recommended lints to
+  # encourage good coding practices. The lint set provided by the package is
+  # activated in the `analysis_options.yaml` file located at the root of your
+  # package. See that file for information about deactivating specific lint
+  # rules and activating additional ones.
+  flutter_lints: ^2.0.0
